+++ src/main/java/com/infeco/keylease/models/Address.java
@@ -15,6 +15,7 @@ public class Address {
         this.street = addressEntity.getStreet();
         this.additionalAddress = addressEntity.getAdditionalAddress();
         this.zipCode = addressEntity.getZipCode();
+        this.town = addressEntity.getTown();
     }
 
     public String getStreet() {